modificaciones

+++ Controllers/liquidacionCompraControlador.php
@@ -5,55 +5,62 @@ class liquidacionCompraControlador extends liquidacionCompraModelo {
     public function cargar_liquidacioncompra_controlador() {
         //$fileP12 = file_get_contents($_FILES['txtArchivoLQ']['tmp_name']);
         $extFile = pathinfo($_FILES['txtArchivoLQ']['name'], PATHINFO_EXTENSION);
+        $extFile = mb_strtolower($extFile, 'utf-8'); //String cambia las letras a minúsculas; strtoupper pone en mayúsculas
         
-        if($extFile == "pdf"){
-            
-            $nuevoPath = "../Archivos_subidos/liquidacioncompra/";
-                    
-            if (is_dir($nuevoPath)) {
-                chmod($nuevoPath, 0777);
-            } else {
-                mkdir($nuevoPath, 0777, true);
-                chmod($nuevoPath, 0777);
-            }
+        if(strlen($_POST['numeroLC']) == 15){
 
-            $path = "Archivos_subidos/liquidacioncompra/".$_POST['numeroReembolsoLC'] . "." . $extFile;
-                    //$_FILES['txtArchivoLQ']['type'];
+            if($extFile == "pdf"){
 
-            if (move_uploaded_file($_FILES['txtArchivoLQ']['tmp_name'], "../".$path)) { //Cargar archivos
+                $nuevoPath = "../Archivos_subidos/liquidacioncompra/";
 
-                $data = [
-                    'idReembolso' => $_POST['idReembLQ'],
-                    'numero' => $_POST['numeroLC'],
-                    'pathArchivo' => $path,
-                    'estado' => "PENDIENTE",
-                    'idUsuarioModifica' => $_SESSION['Usuario']->id,
-                ];
-
-                $respuesta = liquidacionCompraModelo::cargar_liquidacioncompra_modelo($data);
-
-                if (isset($respuesta) && $respuesta->respuesta == "OK") {
-                    return "<script>swal('','Archivo cargado correctamente.','info')
-                        .then((value) => {
-                                    $(`#btnSearch`).click();
-                                });</script>";
+                if (is_dir($nuevoPath)) {
+                    chmod($nuevoPath, 0777);
+                } else {
+                    mkdir($nuevoPath, 0777, true);
+                    chmod($nuevoPath, 0777);
                 }
-                else if(isset($respuesta)){
-                    //eliimnar el archivo
-                    unlink("../".$path);
-                    return "<script>swal('','Error: ".$respuesta->respuesta."','error');</script>";
+
+                $path = "Archivos_subidos/liquidacioncompra/".$_POST['numeroReembolsoLC'] . "." . $extFile;
+                        //$_FILES['txtArchivoLQ']['type'];
+
+                if (move_uploaded_file($_FILES['txtArchivoLQ']['tmp_name'], "../".$path)) { //Cargar archivos
+
+                    $data = [
+                        'idReembolso' => $_POST['idReembLQ'],
+                        'numero' => $_POST['numeroLC'],
+                        'pathArchivo' => $path,
+                        'estado' => "PENDIENTE",
+                        'idUsuarioModifica' => $_SESSION['Usuario']->id,
+                    ];
+
+                    $respuesta = liquidacionCompraModelo::cargar_liquidacioncompra_modelo($data);
+
+                    if (isset($respuesta) && $respuesta->respuesta == "OK") {
+                        return "<script>swal('','Archivo cargado correctamente.','info')
+                            .then((value) => {
+                                        $(`#btnSearch`).click();
+                                    });</script>";
+                    }
+                    else if(isset($respuesta)){
+                        //eliimnar el archivo
+                        unlink("../".$path);
+                        return "<script>swal('','Error: ".$respuesta->respuesta."','error');</script>";
+                    }
+                    else{
+                        unlink("../".$path);
+                        return "<script>swal('','Error al cargar el archivo y al enviar el correo.','error');</script>";
+                    }
+
                 }
                 else{
-                    unlink("../".$path);
-                    return "<script>swal('','Error al cargar el archivo y al enviar el correo.','error');</script>";
+                    return "<script>swal('','No se pudo guardar el archivo en disco.','error');</script>";
                 }
-
+            }else{
+                return "<script>swal('','Solo se puede cargar archivos de tipo PDF.','warning');</script>";
             }
-            else{
-                return "<script>swal('','No se pudo guardar el archivo en disco.','error');</script>";
-            }
-        }else{
-            return "<script>swal('','Solo se puede cargar archivos de tipo PDF.','warning');</script>";
+        }
+        else{
+            return "<script>swal('','El número de liquidación debe tener 15 dígitos.','warning');</script>";
         }
         
     }
